[CLEANUP]

+++ plugins/org.ietr.preesm.memory/src/org/ietr/preesm/memory/allocation/DeGreefAllocator.java
@@ -110,7 +110,7 @@ public class DeGreefAllocator extends MemoryAllocator {
 
       // Alignment constraint
       int align = -1;
-      final Integer typeSize = (Integer) vertex.getPropertyBean().getValue(MemoryExclusionVertex.TYPE_SIZE, Integer.class);
+      final Integer typeSize = vertex.getPropertyBean().getValue(MemoryExclusionVertex.TYPE_SIZE, Integer.class);
       if (this.alignment == 0) {
         align = typeSize;
       } else if (this.alignment > 0) {
